시험장 필요 감독관 수의 최소값

+++ Algorithm/SW test/13458_시험 감독.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int N;
 int A[1000001];
 int B, C;
-int ans;
+long long ans;
 
 int main()
 {
